+++ src/main/java/com/opentime/smartparking_system/model/vo/ParkingVO.java
@@ -14,5 +14,5 @@ public class ParkingVO {
     private int carId;
     private Timestamp entryTime;
     private Timestamp exitTime;
-    private Long fee;
+    private int fee;
 }
